+++ nodejs-playground/basics/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "nodejs-playground",
+  "version": "1.0.0",
+  "description": "NodeJS Playground",
+  "main": "app.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [
+    "Node",
+    "JS",
+    "Express",
+    "MongoDB",
+    "API",
+    "Backend"
+  ],
+  "author": "Sai Vikas Singamsetty",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.21.2",
+    "morgan": "^1.10.0"
+  },
+  "devDependencies": {
+    "@types/morgan": "^1.9.9",
+    "nodemon": "^3.1.9"
+  }
+}
