+++ nodejs-playground/express/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "nodejs-playground",
+  "name": "expressjs-playground",
   "version": "1.0.0",
-  "description": "NodeJS Playground",
-  "main": "app.js",
+  "description": "ExpressJS Playground",
+  "main": "server.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "nodemon server.js"
   },
   "keywords": [
     "Node",
@@ -18,9 +18,11 @@
   "author": "Sai Vikas Singamsetty",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.21.2"
+    "express": "^4.21.2",
+    "morgan": "^1.10.0"
   },
   "devDependencies": {
+    "@types/morgan": "^1.9.9",
     "nodemon": "^3.1.9"
   }
 }
