+++ nodejs-playground/express/server.js
@@ -1,0 +1,6 @@
+import app from "./app.js";
+const PORT = 3000;
+
+app.listen(PORT, () => {
+  console.log("Server has started");
+});
